Separate lag totals calculation in another file
