+++ src/consumeroffset/kafka_offset_collect_test.go
@@ -1,1 +1,7 @@
 package consumeroffset
+
+import "testing"
+
+func TestCollectOffsetsForConsumerGroup(t *testing.T) {
+
+}
